Support short-circuiting in tsv-filter (retain command argument order via getoptInorder). (#8)
Short-circuiting is added by using getopt_inorder (in common) to process and assemble filter operators in the order on the command line. For example, if a numeric field might have a value like 'none', there was no way to run a numeric test on it without triggering a numeric conversion error when processing 'none'. Now, a test for 'none' can precede the numeric test.